Database.Table.Delta

Description

This module provides a Store for tables using Database.Table.SQLite.Simple.

Documentation

newSqlStore :: (AppendPrimaryKey table, IsTableSql (table :. Col "id" Primary), Show (Row table)) => Proxy table -> Connection -> IO (UpdateStore IO [DeltaDB Int (Row table)])

Construct a UpdateStore from an SQL table.

The unique IDs will be stored in a column "id" at the end of each row in the database table.